requirements.txt
asgiref==3.4.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.9
django-grappelli==2.15.3
django-import-export==2.6.1
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
Pillow==8.4.0
psycopg2==2.9.2
python-environ==0.4.54
pytz==2021.3
PyYAML==6.0
sqlparse==0.4.2
tablib==3.1.0
xlrd==2.0.1
xlwt==1.3.0
